Corrected CityController to pass city as opposed to CityInfo to tie with city.html.
Updated TempAndTime method for getting time to account for timezone.  I wasn't able to work out how to do it with the calendar format, so I changed it to one I was able to get to work.

CityServiceTest had a couple things defaulted in that didn't match the given that weatherService should return (time and country name).  Corrected those.